+++ src/renderer/components/BaseButton.vue
@@ -3,6 +3,7 @@ withDefaults(
 	defineProps<{
 		variant?: 'primary' | 'secondary'
 		type?: 'button' | 'submit'
+		disabled?: boolean
 	}>(),
 	{
 		variant: 'primary',
@@ -12,7 +13,7 @@ withDefaults(
 </script>
 
 <template>
-	<button class="button" :class="`button--${variant}`" :type="type">
+	<button class="button" :class="`button--${variant}`" :disabled="disabled" :type="type">
 		<slot />
 	</button>
 </template>